feat: feature oriented architecture
