+++ src/lib/services/feature-toggle-service.ts
@@ -1,6 +1,5 @@
 import {
     CREATE_FEATURE_STRATEGY,
-    StrategyIds,
     EnvironmentVariantEvent,
     FEATURE_UPDATED,
     FeatureArchivedEvent,
@@ -23,6 +22,7 @@ import {
     IEventStore,
     IFeatureEnvironmentInfo,
     IFeatureEnvironmentStore,
+    IFeatureNaming,
     IFeatureOverview,
     IFeatureStrategy,
     IFeatureTagStore,
@@ -33,29 +33,29 @@ import {
     IProjectStore,
     ISegment,
     IStrategyConfig,
+    IStrategyStore,
     IUnleashConfig,
     IUnleashStores,
     IVariant,
+    PotentiallyStaleOnEvent,
     Saved,
     SKIP_CHANGE_REQUEST,
+    StrategiesOrderChangedEvent,
+    StrategyIds,
     Unsaved,
     WeightType,
-    StrategiesOrderChangedEvent,
-    PotentiallyStaleOnEvent,
-    IStrategyStore,
-    IFeatureNaming,
 } from '../types';
 import { Logger } from '../logger';
-import { PatternError } from '../error';
+import {
+    ForbiddenError,
+    FOREIGN_KEY_VIOLATION,
+    OperationDeniedError,
+    PatternError,
+    PermissionError,
+} from '../error';
 import BadDataError from '../error/bad-data-error';
 import NameExistsError from '../error/name-exists-error';
 import InvalidOperationError from '../error/invalid-operation-error';
-import {
-    FOREIGN_KEY_VIOLATION,
-    OperationDeniedError,
-    PermissionError,
-    ForbiddenError,
-} from '../error';
 import {
     constraintSchema,
     featureMetadataSchema,
@@ -96,7 +96,6 @@ import { ISegmentService } from 'lib/segments/segment-service-interface';
 import { IChangeRequestAccessReadModel } from '../features/change-request-access-service/change-request-access-read-model';
 import { checkFeatureFlagNamesAgainstPattern } from '../features/feature-naming-pattern/feature-naming-validation';
 import { IPrivateProjectChecker } from '../features/private-project/privateProjectCheckerType';
-import { CreateDependentFeatureSchema } from '../openapi';
 
 interface IFeatureContext {
     featureName: string;
@@ -122,15 +121,6 @@ export type FeatureNameCheckResultWithFeaturePattern =
           invalidNames: Set<string>;
           featureNaming: IFeatureNaming;
       };
-
-export type FeatureDependency =
-    | {
-          parent: string;
-          child: string;
-          enabled: true;
-          variants?: string[];
-      }
-    | { parent: string; child: string; enabled: false };
 
 const oneOf = (values: string[], match: string) => {
     return values.some((value) => value === match);
@@ -2210,27 +2200,6 @@ class FeatureToggleService {
                 ),
             );
         }
-    }
-
-    async upsertFeatureDependency(
-        parentFeature: string,
-        dependentFeature: CreateDependentFeatureSchema,
-    ): Promise<void> {
-        const { enabled, feature, variants } = dependentFeature;
-        const featureDependency: FeatureDependency =
-            enabled === false
-                ? {
-                      parent: parentFeature,
-                      child: feature,
-                      enabled,
-                  }
-                : {
-                      parent: parentFeature,
-                      child: feature,
-                      enabled: true,
-                      variants,
-                  };
-        console.log(featureDependency);
     }
 }
