+++ src/lib/types/services.ts
@@ -44,6 +44,7 @@ import { ISegmentService } from '../segments/segment-service-interface';
 import ConfigurationRevisionService from '../features/feature-toggle/configuration-revision-service';
 import EventAnnouncerService from 'lib/services/event-announcer-service';
 import { IPrivateProjectChecker } from '../features/private-project/privateProjectCheckerType';
+import { DependentFeaturesService } from '../features/dependent-features/dependent-features-service';
 
 export interface IUnleashServices {
     accessService: AccessService;
@@ -99,4 +100,5 @@ export interface IUnleashServices {
     ) => FeatureToggleService;
     transactionalGroupService: (db: Knex.Transaction) => GroupService;
     privateProjectChecker: IPrivateProjectChecker;
+    dependentFeaturesService: DependentFeaturesService;
 }
